Mais alteracoes de nomes para melhor coesao, implementacao de scoreboard no frontend, correcao de estruturas e remocao de inutilidades (browserrouter) e implementacao inicial do endpoint para definicao de vencedor.
